add dotenv-rails, update DB instructions to use it
To avoid manually passing a password environment variable when invoking
rails commands that use the DB, this introduces dotenv which allows
someone to write a .env file containing environment variables values
they'd like to use. That file is gitignored to avoid causing collisions
with other people's .env file and to avoid leaking anyone's sensitive
local development values. The README is updated advising the person
running the application to use a .env file.