initial react introduction
To be clear, react is not needed for this project, and probably not
needed for many rails projects. But it's a popular framework that people
use, and practicing with it and demonstrating familiarity with it is
valuable. This introduces it via react-rails, and more-or-less recreates
the existing view in the master branch (at least where this branch was
made) using a react component.