make job_positions_controller act as an API, fetch content on page load
Before, rails was rendering a view with the the job position data
injected into the root element as a data attribute, which was passed
along as a property to the JobPositionDashboard component. This pattern
wouldn't scale well as the data attribute on the root element would just
grow continuously. Now, the JobPositionsController has been made into an
API controller that the React component can easily get data from when it
loads.